Refactor strategies
