+++ tests/utils/test_time.py
@@ -6,61 +6,14 @@ import numpy as np
 import pytest
 from cftime import datetime as cftime_datetime
 from hypothesis import given
-from hypothesis import strategies as st
 
+import tests.strategies as pst  # parcels strategies
 from parcels._core.utils.time import (
     TimeInterval,
     _get_cf_attrs,
     maybe_convert_python_timedelta_to_numpy,
     timedelta_to_float,
 )
-
-calendar_strategy = st.sampled_from(
-    [
-        "gregorian",
-        "proleptic_gregorian",
-        "365_day",
-        "360_day",
-        "julian",
-        "366_day",
-        np.datetime64,
-        datetime,
-        np.timedelta64,
-    ]
-)
-
-
-@st.composite
-def np_timedelta64_strategy(draw):
-    """Strategy for generating np.timedelta64 objects."""
-    return np.timedelta64(draw(st.integers(1, 60 * 60 * 24 * 100 * 365)), "s")
-
-
-@st.composite
-def datetime_strategy(draw, calendar=None):
-    if calendar is None:
-        calendar = draw(calendar_strategy)
-    if calendar is np.timedelta64:
-        return draw(np_timedelta64_strategy())
-
-    year = draw(st.integers(1900, 2100))
-    month = draw(st.integers(1, 12))
-    day = draw(st.integers(1, 28))
-    if calendar is datetime:
-        return datetime(year, month, day)
-    if calendar is np.datetime64:
-        return np.datetime64(datetime(year, month, day))
-
-    return cftime_datetime(year, month, day, calendar=calendar)
-
-
-@st.composite
-def time_interval_strategy(draw, left=None, calendar=None):
-    if left is None:
-        left = draw(datetime_strategy(calendar=calendar))
-    right = left + draw(np_timedelta64_strategy())
-
-    return TimeInterval(left, right)
 
 
 @pytest.mark.parametrize(
@@ -83,7 +36,7 @@ def test_time_interval_initialization(left, right):
         TimeInterval(right, left)
 
 
-@given(time_interval_strategy())
+@given(pst.time_interval_strategy())
 def test_time_interval_contains(interval):
     left = 0
     right = timedelta_to_float(interval.right - interval.left)
@@ -94,12 +47,12 @@ def test_time_interval_contains(interval):
     assert interval.is_all_time_in_interval(middle)
 
 
-@given(time_interval_strategy(calendar="365_day"), time_interval_strategy(calendar="365_day"))
+@given(pst.time_interval_strategy(calendar="365_day"), pst.time_interval_strategy(calendar="365_day"))
 def test_time_interval_intersection_commutative(interval1, interval2):
     assert interval1.intersection(interval2) == interval2.intersection(interval1)
 
 
-@given(time_interval_strategy())
+@given(pst.time_interval_strategy())
 def test_time_interval_intersection_with_self(interval):
     assert interval.intersection(interval) == interval
 
@@ -111,7 +64,7 @@ def test_time_interval_repr():
     assert repr(interval) == expected
 
 
-@given(time_interval_strategy())
+@given(pst.time_interval_strategy())
 def test_time_interval_equality(interval):
     assert interval == interval
 
@@ -222,7 +175,7 @@ def test_timedelta_to_float_exceptions():
         timedelta_to_float("invalid_type")
 
 
-@given(datetime_strategy())
+@given(pst.datetime_strategy())
 def test_datetime_get_cf_attrs(dt):
     attrs = _get_cf_attrs(dt)
     assert "seconds" in attrs["units"]
